fix message with links
