+++ src/platform-apps/channels/styles.scss
@@ -337,14 +337,16 @@ $scrollbar-width: 5px;
         }
 
         &-preview {
-          display: inline-block;
+          display: flex;
+          flex-direction: row;
+          align-items: self-start;
           white-space: pre-wrap;
           word-break: break-word;
           font-size: 13.5px;
           font-weight: 500;
           vertical-align: baseline;
           position: relative;
-          max-width: 600px;
+          max-width: 85%;
 
           .text-message {
             display: inline-block;
@@ -352,13 +354,6 @@ $scrollbar-width: 5px;
             padding: 5px 15px;
             margin-top: 10px;
             border-radius: $border-radius;
-          }
-
-          &-with-remove {
-            display: flex;
-            flex-direction: row;
-            align-items: self-start;
-            max-width: 85%;
           }
         }
